chore: Update npm dependencies
This commit updates the npm dependencies in the `package.json` file. It adds the `clsx` package with version `2.1.1`, the `react-icons` package with version `5.2.1`, and the `tailwind-merge` package with version `2.4.0`. It also updates the `next` package to version `14.2.5` and removes the `typescript` package. Additionally, it updates the dev dependencies by adding the `eslint` package with version `8` and the `eslint-config-next` package with version `14.2.5`.