chore: Add Landing component
This commit adds the `Landing` component to the project. The `Landing` component is responsible for rendering the landing page of the application. It takes in various props such as `name`, `role`, `location`, `experienceStart`, `description`, `chipSections`, `links`, and `image`. The component also imports and uses other components such as `Links`, `BottomBar`, `Image`, and `ChipSection`. This addition enhances the application by providing a visually appealing and informative landing page.